+++ packages/expo-doctor/src/utils/__tests__/doctorConfig.test.ts
@@ -1,53 +1,53 @@
 import {
-  getDirectoryCheckExcludes,
-  getDirectoryCheckEnabled,
-  getDirectoryCheckListUnknownPackagesEnabled,
+  getReactNativeDirectoryCheckExcludes,
+  getReactNativeDirectoryCheckEnabled,
+  getReactNativeDirectoryCheckListUnknownPackagesEnabled,
 } from '../doctorConfig';
 
-describe('getDirectoryCheckExcludes', () => {
+describe('getReactNativeDirectoryCheckExcludes', () => {
   it('returns an empty array if no config is present', () => {
-    expect(getDirectoryCheckExcludes({})).toEqual([]);
+    expect(getReactNativeDirectoryCheckExcludes({})).toEqual([]);
   });
 
   it('returns an empty array if the config is empty', () => {
-    expect(getDirectoryCheckExcludes({ expo: { doctor: {} } })).toEqual([]);
+    expect(getReactNativeDirectoryCheckExcludes({ expo: { doctor: {} } })).toEqual([]);
   });
 
   it('returns an empty array if the config has no excludes', () => {
-    expect(getDirectoryCheckExcludes({ expo: { doctor: { directoryCheck: {} } } })).toEqual([]);
+    expect(getReactNativeDirectoryCheckExcludes({ expo: { doctor: { reactNativeDirectoryCheck: {} } } })).toEqual([]);
   });
 
   it('parses strings that begin and end with / as regexes', () => {
     expect(
-      getDirectoryCheckExcludes({
-        expo: { doctor: { directoryCheck: { exclude: ['/foo/', 'bar'] } } },
+      getReactNativeDirectoryCheckExcludes({
+        expo: { doctor: { reactNativeDirectoryCheck: { exclude: ['/foo/', 'bar'] } } },
       })
     ).toEqual([/foo/, 'bar']);
   });
 
   it('returns an array of strings', () => {
     expect(
-      getDirectoryCheckExcludes({
-        expo: { doctor: { directoryCheck: { exclude: ['foo', 'bar'] } } },
+      getReactNativeDirectoryCheckExcludes({
+        expo: { doctor: { reactNativeDirectoryCheck: { exclude: ['foo', 'bar'] } } },
       })
     ).toEqual(['foo', 'bar']);
   });
 });
 
-describe('getDirectoryCheckEnabled', () => {
+describe('getReactNativeDirectoryCheckEnabled', () => {
   it('returns false if the config is empty', () => {
-    expect(getDirectoryCheckEnabled({ expo: { doctor: {} } })).toBe(false);
+    expect(getReactNativeDirectoryCheckEnabled({ expo: { doctor: {} } })).toBe(false);
   });
 
   it('returns true if the config is enabled', () => {
     expect(
-      getDirectoryCheckEnabled({ expo: { doctor: { directoryCheck: { enabled: true } } } })
+      getReactNativeDirectoryCheckEnabled({ expo: { doctor: { reactNativeDirectoryCheck: { enabled: true } } } })
     ).toBe(true);
   });
 
   it('reads from env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK', () => {
     process.env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK = '1';
-    expect(getDirectoryCheckEnabled({ expo: { doctor: {} } })).toBe(true);
+    expect(getReactNativeDirectoryCheckEnabled({ expo: { doctor: {} } })).toBe(true);
     delete process.env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK;
   });
 
@@ -57,35 +57,35 @@ describe('getDirectoryCheckEnabled', () => {
 
     process.env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK = '0';
     expect(
-      getDirectoryCheckEnabled({ expo: { doctor: { directoryCheck: { enabled: true } } } })
+      getReactNativeDirectoryCheckEnabled({ expo: { doctor: { reactNativeDirectoryCheck: { enabled: true } } } })
     ).toBe(false);
     delete process.env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK;
 
     expect(console.warn).toHaveBeenCalledWith(
-      'Both EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK and config.directoryCheck.enabled are set. Using EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK.'
+      'Both EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK and config.reactNativeDirectoryCheck.enabled are set. Using EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK.'
     );
 
     console.warn = originalConsoleWarn;
   });
 });
 
-describe('getDirectoryCheckListUnknownPackagesEnabled', () => {
+describe('getReactNativeDirectoryCheckListUnknownPackagesEnabled', () => {
   it('returns true if the config is empty', () => {
-    expect(getDirectoryCheckListUnknownPackagesEnabled({ expo: { doctor: {} } })).toBe(true);
+    expect(getReactNativeDirectoryCheckListUnknownPackagesEnabled({ expo: { doctor: {} } })).toBe(true);
   });
 
   it('returns true if the config is enabled', () => {
     expect(
-      getDirectoryCheckListUnknownPackagesEnabled({
-        expo: { doctor: { directoryCheck: { listUnknownPackages: true } } },
+      getReactNativeDirectoryCheckListUnknownPackagesEnabled({
+        expo: { doctor: { reactNativeDirectoryCheck: { listUnknownPackages: true } } },
       })
     ).toBe(true);
   });
 
   it('returns false if the config is disabled', () => {
     expect(
-      getDirectoryCheckListUnknownPackagesEnabled({
-        expo: { doctor: { directoryCheck: { listUnknownPackages: false } } },
+      getReactNativeDirectoryCheckListUnknownPackagesEnabled({
+        expo: { doctor: { reactNativeDirectoryCheck: { listUnknownPackages: false } } },
       })
     ).toBe(false);
   });