+++ packages/expo-doctor/CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### 💡 Others
 
-- Rename `directoryCheck` to `reactNativeDirectoryCheck`.
+- Rename `directoryCheck` to `reactNativeDirectoryCheck`. ([#30647](https://github.com/expo/expo/pull/30647) by [@brentvatne](https://github.com/brentvatne))
 
 ## 1.8.2 — 2024-07-19